[IMP] test_new_api: don't run the test in debug assets
Since we now store debug mode in session, and the session is forced to debug
tests in test mode, this tour was setting the debug mode from tests to assets
thus the test files were not loaded.

Since this debug assets was introduced in a refactoring commit and the tour
does not break without it, this commit remove it so the tour can work with
the new system.