[IMP] web_editor: remove assets from documents
"Document" tab in "Select a media" Dialog was displaying
all the assets files. Now all files containing "assets_"
are excluded from the "Document" tab. This is similar to
the regenerateAssets function in debug_manager that delete
assets based on this criteria.

Line 274 in media.js session.uid was called without the
import, this is fixed.

task-1945874