[FIX] calendar: Recurrent events cause traceback
Create a recurrent event with a reminder and an attendee. Traceback.

The initial `browse` was replaced by a custom `search` that mimic the
browse behavior in 9c4b52000fb. The `search` was retrieving too much
records as it is override in `calendar.event` to also include virtual
ids.

opw-1985498
opw-1985552